restructure the project directory tree
