--- app/src/main/java/vn/edu/hcmuaf/fit/travelapp/activity/MainActivity.java
@@ -1,190 +1,0 @@
-package vn.edu.hcmuaf.fit.travelapp.activity;
-
-import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-import android.widget.ArrayAdapter;
-
-import androidx.activity.EdgeToEdge;
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-import androidx.viewpager2.widget.CompositePageTransformer;
-import androidx.viewpager2.widget.MarginPageTransformer;
-
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ValueEventListener;
-
-import java.util.ArrayList;
-
-import vn.edu.hcmuaf.fit.travelapp.adapter.CategoryAdapter;
-import vn.edu.hcmuaf.fit.travelapp.adapter.PopularAdapter;
-import vn.edu.hcmuaf.fit.travelapp.adapter.RecommendedAdapter;
-import vn.edu.hcmuaf.fit.travelapp.adapter.SliderAdapter;
-import vn.edu.hcmuaf.fit.travelapp.domain.Category;
-import vn.edu.hcmuaf.fit.travelapp.domain.Item;
-import vn.edu.hcmuaf.fit.travelapp.domain.Location;
-import vn.edu.hcmuaf.fit.travelapp.R;
-import vn.edu.hcmuaf.fit.travelapp.databinding.ActivityMainBinding;
-import vn.edu.hcmuaf.fit.travelapp.domain.SliderItem;
-
-public class MainActivity extends AppCompatActivity {
-    private ActivityMainBinding binding;
-    private FirebaseDatabase database;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        EdgeToEdge.enable(this);
-        binding = ActivityMainBinding.inflate(getLayoutInflater());
-        setContentView(binding.getRoot());
-        database = FirebaseDatabase.getInstance();
-
-        initLocation();
-        initBanners();
-        initCategory();
-        initPopular();
-        initRecommended();
-    }
-
-    private void initLocation() {
-        DatabaseReference myref = database.getReference("Location");
-        ArrayList<Location> list = new ArrayList<>();
-        myref.addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
-                if (snapshot.exists()) {
-                    for (DataSnapshot issue : snapshot.getChildren()) {
-                        list.add(issue.getValue(Location.class));
-                    }
-
-                    ArrayAdapter<Location> adapter = new ArrayAdapter<>(MainActivity.this, R.layout.sp_item, list);
-                    adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-                    binding.locationSp.setAdapter(adapter);
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
-
-            }
-        });
-    }
-
-    private void banners(ArrayList<SliderItem> items) {
-        binding.viewPager2.setAdapter(new SliderAdapter(items, binding.viewPager2));
-        binding.viewPager2.setClipToPadding(false);
-        binding.viewPager2.setClipChildren(false);
-        binding.viewPager2.setOffscreenPageLimit(3);
-        binding.viewPager2.getChildAt(0).setOverScrollMode(RecyclerView.OVER_SCROLL_NEVER);
-
-        CompositePageTransformer compositePageTransformer = new CompositePageTransformer();
-        compositePageTransformer.addTransformer(new MarginPageTransformer(40));
-        binding.viewPager2.setPageTransformer(compositePageTransformer);
-    }
-
-    private void initBanners() {
-        DatabaseReference myRef = database.getReference("Banner");
-        binding.progressBarBanner.setVisibility(View.VISIBLE);
-        ArrayList<SliderItem> items = new ArrayList<>();
-        myRef.addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
-                if (snapshot.exists()) {
-                    for (DataSnapshot issue : snapshot.getChildren()) {
-                        items.add(issue.getValue(SliderItem.class));
-                    }
-                    banners(items);
-                    binding.progressBarBanner.setVisibility(View.GONE);
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
-
-            }
-        });
-    }
-
-    private void initCategory() {
-        DatabaseReference myref = database.getReference("Category");
-        ArrayList<Category> list = new ArrayList<>();
-        myref.addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
-                if (snapshot.exists()) {
-                    for (DataSnapshot issue : snapshot.getChildren()) {
-                        list.add(issue.getValue(Category.class));
-                    }
-                    if (!list.isEmpty()) {
-                        binding.recyclerViewCategory.setLayoutManager(new LinearLayoutManager(MainActivity.this, LinearLayoutManager.HORIZONTAL, false));
-                        RecyclerView.Adapter adapter = new CategoryAdapter(list);
-                        binding.recyclerViewCategory.setAdapter(adapter);
-                    }
-                    binding.progressBarCategory.setVisibility(View.GONE);
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
-
-            }
-        });
-    }
-
-    private void initPopular() {
-        DatabaseReference myref = database.getReference("Popular");
-        ArrayList<Item> list = new ArrayList<>();
-        myref.addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
-                if (snapshot.exists()) {
-                    for (DataSnapshot issue : snapshot.getChildren()) {
-                        list.add(issue.getValue(Item.class));
-                    }
-                    if (!list.isEmpty()) {
-                        binding.recyclerViewPopular.setLayoutManager(new LinearLayoutManager(MainActivity.this, LinearLayoutManager.HORIZONTAL, false));
-                        RecyclerView.Adapter adapter = new PopularAdapter(list);
-                        binding.recyclerViewPopular.setAdapter(adapter);
-                    }
-                    binding.progressBarPopular.setVisibility(View.GONE);
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
-
-            }
-        });
-    }
-
-    private void initRecommended() {
-        DatabaseReference myref = database.getReference("Item");
-        ArrayList<Item> list = new ArrayList<>();
-        myref.addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
-                if (snapshot.exists()) {
-                    for (DataSnapshot issue : snapshot.getChildren()) {
-                        list.add(issue.getValue(Item.class));
-                    }
-                    if (!list.isEmpty()) {
-                        binding.recyclerViewRecommended.setLayoutManager(new LinearLayoutManager(MainActivity.this, LinearLayoutManager.HORIZONTAL, false));
-                        RecyclerView.Adapter adapter = new RecommendedAdapter(list);
-                        binding.recyclerViewRecommended.setAdapter(adapter);
-                    }
-                    binding.progressBarRecommended.setVisibility(View.GONE);
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
-
-            }
-        });
-    }
-}
