--- app/src/main/java/vn/edu/hcmuaf/fit/travelapp/activity/SplashActivity.java
@@ -1,25 +1,0 @@
-package vn.edu.hcmuaf.fit.travelapp.activity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-
-import androidx.activity.EdgeToEdge;
-import androidx.appcompat.app.AppCompatActivity;
-
-import vn.edu.hcmuaf.fit.travelapp.databinding.ActivitySplashBinding;
-
-public class SplashActivity extends AppCompatActivity {
-    private ActivitySplashBinding binding;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        EdgeToEdge.enable(this);
-        binding = ActivitySplashBinding.inflate(getLayoutInflater());
-        setContentView(binding.getRoot());
-
-        binding.startBtn.setOnClickListener(v -> startActivity(new Intent(SplashActivity.this, MainActivity.class)));
-    }
-}
