--- app/src/main/java/vn/edu/hcmuaf/fit/travelapp/domain/Category.java
@@ -1,34 +1,0 @@
-package vn.edu.hcmuaf.fit.travelapp.domain;
-
-public class Category {
-    private int Id;
-    private String ImagePath;
-    private String Name;
-
-    public Category() {
-    }
-
-    public int getId() {
-        return Id;
-    }
-
-    public void setId(int id) {
-        Id = id;
-    }
-
-    public String getImagePath() {
-        return ImagePath;
-    }
-
-    public void setImagePath(String imagePath) {
-        ImagePath = imagePath;
-    }
-
-    public String getName() {
-        return Name;
-    }
-
-    public void setName(String name) {
-        Name = name;
-    }
-}
