--- app/src/main/java/vn/edu/hcmuaf/fit/travelapp/domain/Location.java
@@ -1,30 +1,0 @@
-package vn.edu.hcmuaf.fit.travelapp.domain;
-
-public class Location {
-    private int Id;
-    private String loc;
-
-    public Location() {
-    }
-
-    public int getId() {
-        return Id;
-    }
-
-    public void setId(int id) {
-        Id = id;
-    }
-
-    public String getLoc() {
-        return loc;
-    }
-
-    public void setLoc(String loc) {
-        this.loc = loc;
-    }
-
-    @Override
-    public String toString() {
-        return loc;
-    }
-}
