--- app/src/main/java/vn/edu/hcmuaf/fit/travelapp/domain/SliderItem.java
@@ -1,16 +1,0 @@
-package vn.edu.hcmuaf.fit.travelapp.domain;
-
-public class SliderItem {
-    private String url;
-
-    public SliderItem() {
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-}
