--- app/src/main/java/vn/edu/hcmuaf/fit/travelapp/product/data/datasource/DeleteCallback.java
@@ -1,6 +1,0 @@
-package vn.edu.hcmuaf.fit.travelapp.product.data.datasource;
-
-public interface DeleteCallback {
-    void onSuccess();
-    void onFailure(Throwable error);
-}
