--- app/src/main/java/vn/edu/hcmuaf/fit/travelapp/product/data/datasource/UploadCallback.java
@@ -1,6 +1,0 @@
-package vn.edu.hcmuaf.fit.travelapp.product.data.datasource;
-
-public interface UploadCallback {
-    void onSuccess(String imageUrl);
-    void onFailure(Throwable error);
-}
