--- app/src/main/java/vn/edu/hcmuaf/fit/travelapp/product/data/model/CloudinaryDeleteResponse.java
@@ -1,9 +1,0 @@
-package vn.edu.hcmuaf.fit.travelapp.product.data.model;
-
-public class CloudinaryDeleteResponse {
-    private String result;
-
-    public String getResult() {
-        return result;
-    }
-}
