--- app/src/main/java/vn/edu/hcmuaf/fit/travelapp/product/data/repository/OnProductActionListener.java
@@ -1,8 +1,0 @@
-package vn.edu.hcmuaf.fit.travelapp.product.data.repository;
-
-import vn.edu.hcmuaf.fit.travelapp.product.data.model.Product;
-
-public interface OnProductActionListener {
-    void onDeleteProduct(Product product);
-    void onUpdateProduct(Product product);
-}
