+++ app/src/main/java/vn/edu/hcmuaf/fit/travelapp/product/home/adapter/CategoryAdapter.java
@@ -1,0 +1,80 @@
+package vn.edu.hcmuaf.fit.travelapp.product.home.adapter;
+
+import android.content.Context;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.bumptech.glide.Glide;
+
+import java.util.List;
+
+import vn.edu.hcmuaf.fit.travelapp.R;
+import vn.edu.hcmuaf.fit.travelapp.databinding.ViewholderCategoryBinding;
+import vn.edu.hcmuaf.fit.travelapp.product.home.data.model.Category;
+
+public class CategoryAdapter extends RecyclerView.Adapter<CategoryAdapter.Viewholder> {
+    private final List<Category> items;
+    private int selectedPosition = -1;
+    private int lastSelectedPosition = -1;
+    private Context context;
+
+    public CategoryAdapter(List<Category> items) {
+        this.items = items;
+    }
+
+    @NonNull
+    @Override
+    public CategoryAdapter.Viewholder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+        context = parent.getContext();
+        ViewholderCategoryBinding binding = ViewholderCategoryBinding.inflate(LayoutInflater.from(context), parent, false);
+        return new Viewholder(binding);
+    }
+
+    @Override
+    public void onBindViewHolder(@NonNull CategoryAdapter.Viewholder holder, int position) {
+        Category item = items.get(position);
+        holder.binding.titleTxt.setText(item.getName());
+
+        Glide.with(holder.itemView.getContext()).load(item.getImagePath()).into(holder.binding.pic);
+
+        holder.binding.getRoot().setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                lastSelectedPosition = selectedPosition;
+                selectedPosition = position;
+                notifyItemChanged(lastSelectedPosition);
+                notifyItemChanged(selectedPosition);
+            }
+        });
+
+        holder.binding.titleTxt.setTextColor(context.getResources().getColor(R.color.white));
+
+        if (selectedPosition==position) {
+            holder.binding.pic.setBackgroundResource(0);
+            holder.binding.mainLayout.setBackgroundResource(R.drawable.purple_bg);
+            holder.binding.titleTxt.setVisibility(View.VISIBLE);
+        } else {
+            holder.binding.pic.setBackgroundResource(R.drawable.grey_bg);
+            holder.binding.mainLayout.setBackgroundResource(0);
+            holder.binding.titleTxt.setVisibility(View.GONE);
+        }
+    }
+
+    @Override
+    public int getItemCount() {
+        return items.size();
+    }
+
+    public class Viewholder extends RecyclerView.ViewHolder {
+        private final ViewholderCategoryBinding binding;
+
+        public Viewholder(ViewholderCategoryBinding binding) {
+            super(binding.getRoot());
+            this.binding = binding;
+        }
+    }
+}
