+++ app/src/main/java/vn/edu/hcmuaf/fit/travelapp/product/home/data/model/SliderItem.java
@@ -1,0 +1,16 @@
+package vn.edu.hcmuaf.fit.travelapp.product.home.data.model;
+
+public class SliderItem {
+    private String url;
+
+    public SliderItem() {
+    }
+
+    public String getUrl() {
+        return url;
+    }
+
+    public void setUrl(String url) {
+        this.url = url;
+    }
+}
