+++ app/src/main/java/vn/edu/hcmuaf/fit/travelapp/product/home/ui/SplashActivity.java
@@ -1,0 +1,23 @@
+package vn.edu.hcmuaf.fit.travelapp.product.home.ui;
+
+import android.content.Intent;
+import android.os.Bundle;
+
+import androidx.activity.EdgeToEdge;
+import androidx.appcompat.app.AppCompatActivity;
+
+import vn.edu.hcmuaf.fit.travelapp.databinding.ActivitySplashBinding;
+
+public class SplashActivity extends AppCompatActivity {
+    private ActivitySplashBinding binding;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        EdgeToEdge.enable(this);
+        binding = ActivitySplashBinding.inflate(getLayoutInflater());
+        setContentView(binding.getRoot());
+
+        binding.startBtn.setOnClickListener(v -> startActivity(new Intent(SplashActivity.this, MainActivity.class)));
+    }
+}
