+++ app/src/main/java/vn/edu/hcmuaf/fit/travelapp/product/home/ui/TicketActivity.java
@@ -1,0 +1,63 @@
+package vn.edu.hcmuaf.fit.travelapp.product.home.ui;
+
+import android.content.Intent;
+import android.net.Uri;
+import android.os.Bundle;
+
+import androidx.activity.EdgeToEdge;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.bumptech.glide.Glide;
+
+import vn.edu.hcmuaf.fit.travelapp.databinding.ActivityTicketBinding;
+import vn.edu.hcmuaf.fit.travelapp.product.home.data.model.Item;
+
+public class TicketActivity extends AppCompatActivity {
+    ActivityTicketBinding binding;
+    private Item object;
+    
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        EdgeToEdge.enable(this);
+        binding = ActivityTicketBinding.inflate(getLayoutInflater());
+        setContentView(binding.getRoot());
+        
+        getIntentExtra();
+        setVariable();
+    }
+
+    private void setVariable() {
+        Glide.with(TicketActivity.this)
+                .load(object.getPic())
+                .into(binding.pic);
+
+        Glide.with(TicketActivity.this)
+                .load(object.getTourGuidePic())
+                .into(binding.profile);
+
+        binding.backBtn.setOnClickListener(v -> finish());
+        binding.titleTxt.setText(object.getTitle());
+        binding.durationTxt.setText(object.getDuration());
+        binding.tourGuideNameTxt.setText(object.getTourGuideName());
+        binding.timeTxt.setText(object.getTimeTour());
+        binding.tourGuideTxt.setText(object.getDateTour());
+
+        binding.callBtn.setOnClickListener(v -> {
+            Intent sendIntent = new Intent(Intent.ACTION_VIEW);
+            sendIntent.setData(Uri.parse("sms:" + object.getTourGuidePhone()));
+            sendIntent.putExtra("sms_body", "type your message");
+            startActivity(sendIntent);
+        });
+
+        binding.messageBtn.setOnClickListener(v -> {
+            String phone = object.getTourGuidePhone();
+            Intent intent = new Intent(Intent.ACTION_DIAL, Uri.fromParts("tel", phone, null));
+            startActivity(intent);
+        });
+    }
+
+    private void getIntentExtra() {
+        object = (Item) getIntent().getSerializableExtra("object");
+    }
+}
