+++ app/src/main/java/vn/edu/hcmuaf/fit/travelapp/product/productManagement/data/datasource/DeleteCallback.java
@@ -1,0 +1,6 @@
+package vn.edu.hcmuaf.fit.travelapp.product.productManagement.data.datasource;
+
+public interface DeleteCallback {
+    void onSuccess();
+    void onFailure(Throwable error);
+}
