+++ app/src/main/java/vn/edu/hcmuaf/fit/travelapp/product/productManagement/data/datasource/UploadCallback.java
@@ -1,0 +1,6 @@
+package vn.edu.hcmuaf.fit.travelapp.product.productManagement.data.datasource;
+
+public interface UploadCallback {
+    void onSuccess(String imageUrl);
+    void onFailure(Throwable error);
+}
