+++ app/src/main/java/vn/edu/hcmuaf/fit/travelapp/product/productManagement/data/model/CloudinaryDeleteResponse.java
@@ -1,0 +1,9 @@
+package vn.edu.hcmuaf.fit.travelapp.product.productManagement.data.model;
+
+public class CloudinaryDeleteResponse {
+    private String result;
+
+    public String getResult() {
+        return result;
+    }
+}
