+++ app/src/main/res/layout/activity_detail.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".activity.DetailActivity">
+    tools:context=".product.home.ui.DetailActivity">
 
     <ScrollView
         android:id="@+id/scrollView3"
