+++ app/src/main/res/layout/activity_splash.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.SplashActivity">
+    tools:context=".product.home.ui.SplashActivity">
 
     <ImageView
         android:id="@+id/imageView"
