+++ app/src/main/res/layout/activity_ticket.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/lightGray"
-    tools:context=".activity.TicketActivity">
+    tools:context=".product.home.ui.TicketActivity">
 
     <ScrollView
         android:layout_width="match_parent"
